🦀first pass to deploy content
