--- gradio-app/README.md
@@ -1,1 +1,0 @@
-Until we found a better place..
