+++ gradio-app/apps/__init__.py
@@ -1,0 +1,5 @@
+"""
+This package contains different Gradio app interfaces:
+- chat_interface: Simple chat interface for LLM interaction
+- prompt_playground: Advanced prompt testing interface with history
+"""
