--- gradio-app/chart/Chart.yaml
@@ -1,10 +1,0 @@
-apiVersion: v2
-name: gradio-prompt-app
-description: A Helm chart to deploy Gradio Prompt app 
-type: application
-version: 1.0.0
-appVersion: 1.0.0 
-maintainers:
-  - name: rcarrata
-  - name: rhrolun
-  - name: ckavili
