--- gradio-app/chart/templates/deployment.yaml
@@ -1,49 +1,0 @@
----
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: gradio-prompt-app
-  labels:
-    app: gradio-prompt-app
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: gradio-prompt-app
-  template:
-    metadata:
-      labels:
-        app: gradio-prompt-app
-    spec:
-      containers:
-      - name: gradio-prompt-app
-        image: quay.io/rhoai-genaiops/gradio-prompt-app:0.0.1
-        imagePullPolicy: Always
-        ports:
-        - containerPort: 7860
-          name: http
-          protocol: TCP
-        resources:
-          requests:
-            memory: "256Mi"
-            cpu: "250m"
-          limits:
-            memory: "512Mi"
-            cpu: "500m"
-        livenessProbe:
-          httpGet:
-            path: /
-            port: 7860
-          initialDelaySeconds: 30
-          periodSeconds: 10
-          timeoutSeconds: 5
-          failureThreshold: 3
-        readinessProbe:
-          httpGet:
-            path: /
-            port: 7860
-          initialDelaySeconds: 5
-          periodSeconds: 5
-          timeoutSeconds: 3
-          failureThreshold: 3
-      restartPolicy: Always
