+++ gradio-app/requirements.txt
@@ -1,2 +1,5 @@
 gradio>=4.0.0
 requests>=2.31.0
+fastapi>=0.104.0
+uvicorn>=0.24.0
+python-multipart>=0.0.6
