+++ install.sh
@@ -1,16 +1,18 @@
-git clone https://github.com/rhoai-mlops/deploy-lab.git
 
-## Install GitOps Operator and Instance
-helm upgrade --install ml500-gitops gitops --namespace ml500 --create-namespace
+# Install the operators
+helm upgrade --install ai501-base operators --namespace ai501 --create-namespace
 
-## Install Operators (this will be in an AppSet)
-helm upgrade --install ml500-operators operators --namespace ml500 --create-namespace
+oc wait --for=jsonpath='{.status.availableReplicas}'=1 -n openshift-gitops deployment/cluster
 
-## Install helm chart (this will be in an AppSet)
-helm upgrade --install ml500-base charts/ --namespace ml500 --create-namespace
+# Install the toolings
+helm upgrade --install ai501-toolings toolings --namespace ai501 --create-namespace
 
-# Patch OAuth to point to ML500 htpasswd
-oc patch --type=merge OAuth/cluster -p '{"spec": {"identityProviders": [{"name": "Students", "type": "HTPasswd", "mappingMethod": "claim", "htpasswd": {"fileData": {"name": "htpasswd-ml500"}}}, {"name": "htpasswd_provider", "type": "HTPasswd", "mappingMethod": "claim", "htpasswd": {"fileData": {"name": "htpasswd"}}}]}}'
+# Install the toolings
+helm upgrade --install ai501-student-content student-content --namespace ai501 --create-namespace
+
+
+# Patch OAuth to point to ai501 htpasswd
+oc patch --type=merge OAuth/cluster -p '{"spec": {"identityProviders": [{"name": "Students", "type": "HTPasswd", "mappingMethod": "claim", "htpasswd": {"fileData": {"name": "htpasswd-ai501"}}}, {"name": "htpasswd_provider", "type": "HTPasswd", "mappingMethod": "claim", "htpasswd": {"fileData": {"name": "htpasswd"}}}]}}'
 
 oc patch --type=merge DataScienceCluster/default-dsc -p '{"spec": {"components": {"trustyai": {"managementState": "Managed", "devFlags": {"manifests": [{"contextDir": "config", "sourcePath": "", "uri": "https://api.github.com/repos/ckavili/trustyai-service-operator-ci/tarball/service-acca8f52f3f163444b2fc68003af5cae13f04762"}]}}}}}'
 
@@ -19,7 +21,7 @@ oc -n openshift-user-workload-monitoring patch configmap user-workload-monitorin
 oc patch config.imageregistry.operator.openshift.io/cluster --patch '{"spec":{"defaultRoute":true}}' --type=merge
 
 # Make ArgoCD cluster wide
-attendees=`grep attendees charts/values.yaml | cut -d':' -f2`
+attendees=`grep attendees student-content/values.yaml | cut -d':' -f2`
  for ((i=0; i<=$attendees; i++))
  do
    if [ $i -eq 1 ]; then
