+++ operators/templates/acs.yaml
@@ -1,27 +1,27 @@
----
-apiVersion: v1
-kind: Namespace
-metadata:
-  name: rhacs-operator
----
-apiVersion: operators.coreos.com/v1
-kind: OperatorGroup
-metadata:
-  name: rhacs-operator
-  namespace: rhacs-operator
-spec:
-  upgradeStrategy: Default
----
-apiVersion: operators.coreos.com/v1alpha1
-kind: Subscription
-metadata:
-  labels:
-    operators.coreos.com/rhacs-operator.rhacs-operator: ''
-  name: rhacs-operator
-  namespace: rhacs-operator
-spec:
-  channel: stable
-  installPlanApproval: Automatic
-  name: rhacs-operator
-  source: redhat-operators
-  sourceNamespace: openshift-marketplace
+# ---
+# apiVersion: v1
+# kind: Namespace
+# metadata:
+#   name: rhacs-operator
+# ---
+# apiVersion: operators.coreos.com/v1
+# kind: OperatorGroup
+# metadata:
+#   name: rhacs-operator
+#   namespace: rhacs-operator
+# spec:
+#   upgradeStrategy: Default
+# ---
+# apiVersion: operators.coreos.com/v1alpha1
+# kind: Subscription
+# metadata:
+#   labels:
+#     operators.coreos.com/rhacs-operator.rhacs-operator: ''
+#   name: rhacs-operator
+#   namespace: rhacs-operator
+# spec:
+#   channel: stable
+#   installPlanApproval: Automatic
+#   name: rhacs-operator
+#   source: redhat-operators
+#   sourceNamespace: openshift-marketplace
