+++ operators/templates/namespace.yaml
@@ -2,4 +2,4 @@
 apiVersion: v1
 kind: Namespace
 metadata:
-  name: sealed-secrets
+  # name: sealed-secrets
