+++ student-content/templates/datascienceproject/dspa.yaml
@@ -9,7 +9,7 @@ metadata:
     "helm.sh/hook": post-install,post-upgrade
     "helm.sh/hook-weight": "3"
   name: dspa
-  namespace: user{{ $attendee }}-jukebox
+  namespace: user{{ $attendee }}-canopy
   finalizers:
     - datasciencepipelinesapplications.opendatahub.io/finalizer
 spec:
@@ -34,7 +34,7 @@ spec:
     externalStorage:
       basePath: ''
       bucket: pipeline
-      host: 'minio-api-user{{ $attendee }}-jukebox.{{ $cluster_domain }}'
+      host: 'minio-api-user{{ $attendee }}-canopy.{{ $cluster_domain }}'
       port: ''
       region: wonderland
       s3CredentialsSecret:
