+++ student-content/templates/datascienceproject/namespace.yaml
@@ -4,16 +4,15 @@
 kind: Namespace
 apiVersion: v1
 metadata:
-  name: user{{ $attendee}}-jukebox
+  name: user{{ $attendee}}-canopy
   labels:
-    kubernetes.io/metadata.name: user{{ $attendee}}-jukebox
+    kubernetes.io/metadata.name: user{{ $attendee}}-canopy
     opendatahub.io/dashboard: 'true'
-    openshift-pipelines.tekton.dev/namespace-reconcile-version: 1.15.1
     maistra.io/member-of: istio-system
     modelmesh-enabled: 'false'
   annotations:
-    openshift.io/description: 'user{{ $attendee}} Jukebox Experiment Environment'
-    openshift.io/display-name: user{{ $attendee}}-jukebox
+    openshift.io/description: 'user{{ $attendee}} Canopy Experiment Environment'
+    openshift.io/display-name: user{{ $attendee}}-canopy
     openshift.io/requester: user{{ $attendee }}
 spec:
   finalizers:
