+++ student-content/templates/datascienceproject/pv-models.yaml
@@ -4,8 +4,8 @@
 apiVersion: v1
 kind: PersistentVolumeClaim
 metadata:
-  name: jukebox-model-pvc
-  namespace: user{{ $attendee }}-jukebox
+  name: canopy-model-pvc
+  namespace: user{{ $attendee }}-canopy
   annotations:
     volume.beta.kubernetes.io/storage-provisioner: kubernetes.io/aws-ebs
 spec:
