+++ student-content/templates/datascienceproject/rolebindings.yaml
@@ -5,7 +5,7 @@ kind: RoleBinding
 apiVersion: rbac.authorization.k8s.io/v1
 metadata:
   name: admin
-  namespace: user{{ $attendee }}-jukebox
+  namespace: user{{ $attendee }}-canopy
 subjects:
   - kind: User
     apiGroup: rbac.authorization.k8s.io
