--- student-content/templates/feast-database/deployment.yaml
@@ -1,78 +1,0 @@
-{{ $attendees := (.Values.attendees | int) }}
-{{- range $attendee := until $attendees }}
----
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: feature-store
-  namespace: user{{ $attendee }}-jukebox
-spec:
-  selector:
-    matchLabels:
-      app: feature-store
-  replicas: 1
-  template:
-    metadata:
-      labels:
-        app: feature-store
-    spec:
-      containers:
-        - name: postgresql
-          image: registry.redhat.io/rhel9/postgresql-13:latest
-          resources:
-            limits:
-              memory: 512Mi
-          readinessProbe:
-            exec:
-              command:
-                - /usr/libexec/check-container
-            initialDelaySeconds: 5
-            timeoutSeconds: 1
-            periodSeconds: 10
-            successThreshold: 1
-            failureThreshold: 3
-          livenessProbe:
-            exec:
-              command:
-                - /usr/libexec/check-container
-                - '--live'
-            initialDelaySeconds: 120
-            timeoutSeconds: 10
-            periodSeconds: 10
-            successThreshold: 1
-            failureThreshold: 3
-          env:
-            - name: POSTGRESQL_USER
-              valueFrom:
-                secretKeyRef:
-                  name: feast
-                  key: database-user
-            - name: POSTGRESQL_PASSWORD
-              valueFrom:
-                secretKeyRef:
-                  name: feast
-                  key: database-password
-            - name: POSTGRESQL_DATABASE
-              valueFrom:
-                secretKeyRef:
-                  name: feast
-                  key: database-name
-          securityContext:
-            capabilities: {}
-            privileged: false
-          ports:
-            - containerPort: 5432
-              protocol: TCP
-          imagePullPolicy: IfNotPresent
-          terminationMessagePath: /dev/termination-log
-          terminationMessagePolicy: File
-          volumeMounts:
-            - name: feast-data
-              mountPath: /var/lib/pgsql/data
-      volumes:
-        - name: feast-data
-          persistentVolumeClaim:
-            claimName: feast
-  strategy:
-    type: Recreate
-{{- end }}
