--- student-content/templates/feast-database/pvc.yaml
@@ -1,19 +1,0 @@
-{{ $attendees := (.Values.attendees | int) }}
-{{- range $attendee := until $attendees }}
----
-apiVersion: v1
-kind: PersistentVolumeClaim
-metadata:
-  name: feast
-  namespace: user{{ $attendee }}-jukebox
-  labels:
-    app: feature-store
-  annotations:
-    argocd.argoproj.io/sync-wave: "1"
-spec:
-  accessModes:
-  - ReadWriteOnce
-  resources:
-    requests:
-      storage: 10Gi
-{{- end }}
