--- student-content/templates/feast-database/secret.yaml
@@ -1,16 +1,0 @@
-{{ $attendees := (.Values.attendees | int) }}
-{{- range $attendee := until $attendees }}
----
-kind: Secret
-apiVersion: v1
-metadata:
-  name: feast
-  namespace: user{{ $attendee }}-jukebox
-  labels:
-    app: feature-store
-stringData:
-  database-name: feast
-  database-password: feast
-  database-user: feast
-type: Opaque
-{{- end }}
