+++ student-content/templates/image-puller/image-puller.yaml
@@ -6,4 +6,4 @@ metadata:
   namespace: openshift-operators
 spec:
   daemonsetName: k8s-image-puller
-  images: 'ds=quay.io/modh/odh-generic-data-science-notebook@sha256:36454fcf796ea284b1e551469fa1705d64c4b2698a8bf9df82a84077989faf5e;wb=image-registry.openshift-image-registry.svc:5000/redhat-ods-applications/s2i-generic-data-science-notebook:2024.2;trusty=image-registry.openshift-image-registry.svc:5000/redhat-ods-applications/odh-trustyai-notebook:2024.2;oauth=registry.redhat.io/openshift4/ose-oauth-proxy@sha256:4bef31eb993feb6f1096b51b4876c65a6fb1f4401fee97fa4f4542b6b7c9bc46;code-server=image-registry.openshift-image-registry.svc:5000/redhat-ods-applications/code-server-notebook:2024.2;pipeline-1=registry.redhat.io/rhoai/odh-ml-pipelines-driver-rhel8@sha256:161ce73c5f397fea57c0e7ab03fbc36b724c2d49024e1e98da9c9d61aedcfd36;pipeline-2=registry.redhat.io/rhoai/odh-data-science-pipelines-argo-argoexec-rhel8@sha256:bc3e44adbd55d45a0213bf5733f7c9ef08c0738ad431e4f279d9cd6c2044ab44;pipeline-3=quay.io/modh/runtime-images@sha256:1186ac6c9026d1091f707fe8cedfcc1ea12d1ec46edd9e8d56bb4b12ba048630;pipeline-4=registry.redhat.io/rhoai/odh-ml-pipelines-launcher-rhel8@sha256:cd14c7de40df05ddef44ab0da08d1ef1e97e8ae8928f0e466151ede62e5195a2;pipeline-5=registry.redhat.io/openshift-pipelines/pipelines-workingdirinit-rhel8@sha256:aabba024b766391dcec9bd5d5144034f6d5dafbd271fb569edcd50b47c6ddb87;pipeline-6=quay.io/openshift/origin-cli:4.14;helm-task=quay.io/redhat-cop/tekton-task-helm:3.6.3;kfp=quay.io/rlundber/mlops-ct-cicd:0.4;openvion=quay.io/modh/kserve-storage-initializer@sha256:e745f50a8abbe412cdf5d868c5aeb4658f2042837d1d62c4825dac51bd7f40fb;openvino-2=quay.io/modh/openvino_model_server@sha256:e4ece481b8dd866163da7ffcd3e95389ef7fb3d32a9a0dbb056b07839652a4f7;jukebox=quay.io/rhoai-mlops/jukebox:latest;custom-codeserver=quay.io/rhoai-mlops/custom-code-server:0.0.3;'
+  images: 'ds=quay.io/modh/odh-generic-data-science-notebook@sha256:36454fcf796ea284b1e551469fa1705d64c4b2698a8bf9df82a84077989faf5e;wb=image-registry.openshift-image-registry.svc:5000/redhat-ods-applications/s2i-generic-data-science-notebook:2024.2;trusty=image-registry.openshift-image-registry.svc:5000/redhat-ods-applications/odh-trustyai-notebook:2024.2;oauth=registry.redhat.io/openshift4/ose-oauth-proxy@sha256:4bef31eb993feb6f1096b51b4876c65a6fb1f4401fee97fa4f4542b6b7c9bc46;code-server=image-registry.openshift-image-registry.svc:5000/redhat-ods-applications/code-server-notebook:2024.2;pipeline-1=registry.redhat.io/rhoai/odh-ml-pipelines-driver-rhel8@sha256:161ce73c5f397fea57c0e7ab03fbc36b724c2d49024e1e98da9c9d61aedcfd36;pipeline-2=registry.redhat.io/rhoai/odh-data-science-pipelines-argo-argoexec-rhel8@sha256:bc3e44adbd55d45a0213bf5733f7c9ef08c0738ad431e4f279d9cd6c2044ab44;pipeline-3=quay.io/modh/runtime-images@sha256:1186ac6c9026d1091f707fe8cedfcc1ea12d1ec46edd9e8d56bb4b12ba048630;pipeline-4=registry.redhat.io/rhoai/odh-ml-pipelines-launcher-rhel8@sha256:cd14c7de40df05ddef44ab0da08d1ef1e97e8ae8928f0e466151ede62e5195a2;pipeline-5=registry.redhat.io/openshift-pipelines/pipelines-workingdirinit-rhel8@sha256:aabba024b766391dcec9bd5d5144034f6d5dafbd271fb569edcd50b47c6ddb87;pipeline-6=quay.io/openshift/origin-cli:4.14;helm-task=quay.io/redhat-cop/tekton-task-helm:3.6.3;kfp=quay.io/rlundber/mlops-ct-cicd:0.4;openvion=quay.io/modh/kserve-storage-initializer@sha256:e745f50a8abbe412cdf5d868c5aeb4658f2042837d1d62c4825dac51bd7f40fb;openvino-2=quay.io/modh/openvino_model_server@sha256:e4ece481b8dd866163da7ffcd3e95389ef7fb3d32a9a0dbb056b07839652a4f7;canopy=quay.io/rhoai-mlops/canopy:latest;custom-codeserver=quay.io/rhoai-mlops/custom-code-server:0.0.3;'
