+++ student-content/templates/minio/minio.yaml
@@ -5,7 +5,7 @@ kind: PersistentVolumeClaim
 apiVersion: v1
 metadata:
   name: minio-pvc
-  namespace: user{{ $attendee }}-jukebox
+  namespace: user{{ $attendee }}-canopy
 spec:
   accessModes:
     - ReadWriteOnce
@@ -18,7 +18,7 @@ kind: Secret
 apiVersion: v1
 metadata:
   name: minio-secret
-  namespace: user{{ $attendee }}-jukebox
+  namespace: user{{ $attendee }}-canopy
 stringData:
   # change the username and password to your own values.
   # ensure that the user is at least 3 characters long and the password at least 8
@@ -29,7 +29,7 @@ kind: Deployment
 apiVersion: apps/v1
 metadata:
   name: minio
-  namespace: user{{ $attendee }}-jukebox
+  namespace: user{{ $attendee }}-canopy
 spec:
   replicas: 1
   selector:
@@ -114,7 +114,7 @@ kind: Service
 apiVersion: v1
 metadata:
   name: minio-service
-  namespace: user{{ $attendee }}-jukebox
+  namespace: user{{ $attendee }}-canopy
 spec:
   ipFamilies:
     - IPv4
@@ -138,7 +138,7 @@ kind: Route
 apiVersion: route.openshift.io/v1
 metadata:
   name: minio-api
-  namespace: user{{ $attendee }}-jukebox
+  namespace: user{{ $attendee }}-canopy
 spec:
   to:
     kind: Service
@@ -155,7 +155,7 @@ kind: Route
 apiVersion: route.openshift.io/v1
 metadata:
   name: minio-ui
-  namespace: user{{ $attendee }}-jukebox
+  namespace: user{{ $attendee }}-canopy
 spec:
   to:
     kind: Service
