--- student-content/templates/model-registry/model-registry-db-parameters.yaml
@@ -1,13 +1,0 @@
-{{ $attendees := (.Values.attendees | int) }}
-{{- range $attendee := until $attendees }}
----
-kind: ConfigMap
-apiVersion: v1
-metadata:
-  name: user{{ $attendee }}-registry-db-parameters
-  namespace: user{{ $attendee }}-jukebox
-data:
-  MYSQL_ALLOW_EMPTY_PASSWORD: 'true'
-  MYSQL_DATABASE: metadb
-  MYSQL_PORT: '3306'
-{{- end }}
