--- student-content/templates/model-registry/model-registry-secret.yaml
@@ -1,18 +1,0 @@
-{{ $attendees := (.Values.attendees | int) }}
-{{- range $attendee := until $attendees }}
----
-kind: Secret
-apiVersion: v1
-metadata:
-  name: user{{ $attendee }}-registry-db
-  namespace: rhoai-model-registries
-  annotations:
-    template.openshift.io/expose-database_name: '{.data[''database-name'']}'
-    template.openshift.io/expose-password: '{.data[''database-password'']}'
-    template.openshift.io/expose-username: '{.data[''database-user'']}'
-data:
-  database-name: bWV0YWRi
-  database-password: dGVzdA==
-  database-user: cm9vdA==
-type: Opaque
-{{- end }}
