--- student-content/templates/model-registry/modelregistry.yaml
@@ -1,68 +1,0 @@
-{{ $cluster_domain := .Values.cluster_domain }}
-{{ $attendees := (.Values.attendees | int) }}
-{{- range $attendee := until $attendees }}
----
-apiVersion: modelregistry.opendatahub.io/v1alpha1
-kind: ModelRegistry
-metadata:
-  annotations:
-    openshift.io/description: ''
-    openshift.io/display-name: user{{ $attendee }}-registry
-  name: user{{ $attendee }}-registry
-  namespace: rhoai-model-registries
-  finalizers:
-    - modelregistry.opendatahub.io/finalizer
-spec:
-  grpc:
-    image: 'registry.redhat.io/rhoai/odh-mlmd-grpc-server-rhel8@sha256:927a88bade78e121d84e4d072f07a203fb00173af3615e00ce712d45329dde2d'
-    port: 9090
-    resources:
-      limits:
-        cpu: 100m
-        memory: 256Mi
-      requests:
-        cpu: 100m
-        memory: 256Mi
-  istio:
-    audiences:
-      - 'https://kubernetes.default.svc'
-    authConfigLabels:
-      security.opendatahub.io/authorization-group: default
-    authProvider: redhat-ods-applications-auth-provider
-    gateway:
-      domain: {{ $cluster_domain }}
-      grpc:
-        gatewayRoute: enabled
-        port: 443
-        tls:
-          credentialName: default-modelregistry-cert
-          mode: SIMPLE
-      istioIngress: ingressgateway
-      rest:
-        gatewayRoute: enabled
-        port: 443
-        tls:
-          credentialName: default-modelregistry-cert
-          mode: SIMPLE
-    tlsMode: ISTIO_MUTUAL
-  mysql:
-    database: metadb
-    host: user{{ $attendee }}-registry-db.user{{ $attendee }}-jukebox.svc.cluster.local
-    passwordSecret:
-      key: database-password
-      name: user{{ $attendee }}-registry-db
-    port: 3306
-    skipDBCreation: false
-    username: root
-  rest:
-    image: 'quay.io/opendatahub/model-registry:v0.2.10'
-    port: 8080
-    resources:
-      limits:
-        cpu: 100m
-        memory: 256Mi
-      requests:
-        cpu: 100m
-        memory: 256Mi
-    serviceRoute: disabled
-{{- end }}
