--- student-content/templates/model-registry/mysql-pvc.yaml
@@ -1,15 +1,0 @@
-{{ $attendees := (.Values.attendees | int) }}
-{{- range $attendee := until $attendees }}
----
-kind: PersistentVolumeClaim
-apiVersion: v1
-metadata:
-  name: metadata-mysql
-  namespace: user{{ $attendee }}-jukebox
-spec:
-  accessModes:
-    - ReadWriteOnce
-  resources:
-    requests:
-      storage: 10Gi
-{{- end }}
