--- student-content/templates/model-registry/mysql-secret.yaml
@@ -1,13 +1,0 @@
-{{ $attendees := (.Values.attendees | int) }}
-{{- range $attendee := until $attendees }}
----
-kind: Secret
-apiVersion: v1
-metadata:
-  name: user{{ $attendee }}-registry-db-secrets
-  namespace: user{{ $attendee }}-jukebox
-data:
-  MYSQL_ROOT_PASSWORD: dGVzdA==
-  MYSQL_USER_NAME: cm9vdA==
-type: Opaque
-{{- end }}
