--- student-content/templates/model-registry/mysql-service.yaml
@@ -1,20 +1,0 @@
-{{ $attendees := (.Values.attendees | int) }}
-{{- range $attendee := until $attendees }}
----
-kind: Service
-apiVersion: v1
-metadata:
-  name: user{{ $attendee }}-registry-db
-  namespace: user{{ $attendee }}-jukebox
-  labels:
-    component: db
-spec:
-  ports:
-    - name: dbapi
-      protocol: TCP
-      port: 3306
-      targetPort: 3306
-  type: ClusterIP
-  selector:
-    component: db
-{{- end }}
