--- student-content/templates/model-registry/rolebinding.yaml
@@ -1,24 +1,0 @@
-{{ $attendees := (.Values.attendees | int) }}
-{{- range $attendee := until $attendees }}
----
-kind: RoleBinding
-apiVersion: rbac.authorization.k8s.io/v1
-metadata:
-  name: dashboard-permissions-user{{ $attendee }}
-  namespace: rhoai-model-registries
-  labels:
-    app: user{{ $attendee }}-registry
-    app.kubernetes.io/component: user{{ $attendee }}-registry
-    app.kubernetes.io/name: user{{ $attendee }}-registry
-    app.kubernetes.io/part-of: user{{ $attendee }}-registry
-    component: user{{ $attendee }}-registry
-    opendatahub.io/dashboard: 'true'
-subjects:
-  - kind: User
-    apiGroup: rbac.authorization.k8s.io
-    name: user{{ $attendee }}
-roleRef:
-  apiGroup: rbac.authorization.k8s.io
-  kind: Role
-  name: registry-user-user{{ $attendee }}-registry
-{{- end }}
