+++ student-content/values.yaml
@@ -6,5 +6,3 @@ minio:
   buckets:
     - name: pipeline
     - name: models
-    - name: data
-    - name: data-cache
