+++ toolings/Chart.yaml
@@ -1,11 +1,10 @@
 apiVersion: v2
-name: ml500-operators
-description: A Helm chart to deploy the ML500 Toolings 
+name: ai501-toolings
+description: A Helm chart to deploy the AI501 Toolings 
 type: application
 version: 1.0.0
 appVersion: 1.0.0 
 maintainers:
   - name: rcarrata
-  - name: nstrung
   - name: rhrolun
   - name: ckavili
