+++ toolings/templates/datasciencecluster.yaml
@@ -15,9 +15,12 @@ spec:
       managementState: Managed
     kserve:
       managementState: Managed
+      defaultDeploymentMode: Serverless
+      rawDeploymentServiceConfig: Headless
       serving:
         ingressGateway:
           certificate:
+            secretName: knative-serving-cert
             type: OpenshiftDefaultIngress
         managementState: Managed
         name: knative-serving
