+++ toolings/templates/docling-serve/deployment.yaml
@@ -1,3 +1,4 @@
+---
 kind: Deployment
 apiVersion: apps/v1
 metadata:
