+++ toolings/templates/docling-serve/service.yaml
@@ -1,3 +1,4 @@
+---
 kind: Service
 apiVersion: v1
 metadata:
