+++ toolings/templates/gradio/deployment.yaml
@@ -1,0 +1,36 @@
+---
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: gradio-app
+  labels:
+    app: gradio-app
+spec:
+  replicas: 1
+  selector:
+    matchLabels:
+      app: gradio-app
+  template:
+    metadata:
+      labels:
+        app: gradio-app
+    spec:
+      containers:
+      - name: gradio-app
+        image: quay.io/rhoai-genaiops/gradio-app:0.0.1
+        imagePullPolicy: Always
+        env:
+        - name: MODEL_URL
+          value: "{{ .Values.modelUrl | default "" }}"
+        ports:
+        - containerPort: 7860
+          name: http
+          protocol: TCP
+        resources:
+          requests:
+            memory: "256Mi"
+            cpu: "250m"
+          limits:
+            memory: "512Mi"
+            cpu: "500m"
+      restartPolicy: Always
