+++ toolings/templates/gradio/route.yaml
@@ -1,13 +1,17 @@
+---
 kind: Route
 apiVersion: route.openshift.io/v1
+metadata:
+  name: gradio-app
 spec:
   path: /
   to:
     kind: Service
-    name: docling-v0-7-0-predictor
+    name: gradio-app
     weight: 100
   port:
     targetPort: http
   tls:
     termination: edge
+    insecureEdgeTerminationPolicy: Redirect
   wildcardPolicy: None
