+++ toolings/templates/gradio/service.yaml
@@ -2,9 +2,9 @@
 apiVersion: v1
 kind: Service
 metadata:
-  name: gradio-prompt-app
+  name: gradio-app
   labels:
-    app: gradio-prompt-app
+    app: gradio-app
 spec:
   type: ClusterIP
   ports:
@@ -13,4 +13,4 @@ spec:
     protocol: TCP
     name: http
   selector:
-    app: gradio-prompt-app
+    app: gradio-app
